added cart delete() in DeleteUserController added data to InjectDataController fixing naming and url addresses
