+++ src/main/java/online/shop/controller/AddProductController.java
@@ -8,9 +8,8 @@ import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import java.io.IOException;
-import java.util.List;
 
-public class ProductGetAllController extends HttpServlet {
+public class AddProductController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("online.shop");
     private final ProductService productService =
             (ProductService) injector.getInstance(ProductService.class);
@@ -18,8 +17,16 @@ public class ProductGetAllController extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
             throws ServletException, IOException {
-        List<Product> allProducts = productService.getAll();
-        req.setAttribute("products", allProducts);
-        req.getRequestDispatcher("/WEB-INF/views/product/get-all.jsp").forward(req, resp);
+        req.getRequestDispatcher("/WEB-INF/views/product/add.jsp").forward(req, resp);
+    }
+
+    @Override
+    protected void doPost(HttpServletRequest req, HttpServletResponse resp)
+            throws ServletException, IOException {
+        String name = req.getParameter("name");
+        String price = req.getParameter("price");
+        Product product = new Product(name, Double.parseDouble(price));
+        productService.create(product);
+        resp.sendRedirect(req.getContextPath() + "/");
     }
 }
