+++ src/main/java/online/shop/controller/DeleteUserController.java
@@ -1,6 +1,7 @@
 package online.shop.controller;
 
 import online.shop.lib.Injector;
+import online.shop.service.ShoppingCartService;
 import online.shop.service.UserService;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
@@ -11,6 +12,8 @@ import java.io.IOException;
 public class DeleteUserController extends HttpServlet {
     private static final Injector injector = Injector.getInstance("online.shop");
     private final UserService userService = (UserService) injector.getInstance(UserService.class);
+    private final ShoppingCartService shoppingCartService = (ShoppingCartService) injector
+            .getInstance(ShoppingCartService.class);
 
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp)
@@ -18,6 +21,7 @@ public class DeleteUserController extends HttpServlet {
         String userId = req.getParameter("id");
         Long id = Long.valueOf(userId);
         userService.delete(id);
+        shoppingCartService.delete(id);
         resp.sendRedirect(req.getContextPath() + "/users/all");
     }
 }
