+++ src/main/webapp/WEB-INF/views/product/add.jsp
@@ -4,7 +4,7 @@
     <title>Please, put goods! </title>
 </head>
 <body>
-<form method="post" action="${pageContext.request.contextPath}/add">
+<form method="post" action="${pageContext.request.contextPath}/products/add">
     <label for="name">Name:</label><br>
     <input type="text" id="name" name="name"><br>
     <label for="price">Price:</label><br>
