+++ src/main/webapp/web.xml
@@ -19,7 +19,7 @@
     </servlet>
     <servlet-mapping>
         <servlet-name>inject-data</servlet-name>
-        <url-pattern>/injectData</url-pattern>
+        <url-pattern>/inject-data</url-pattern>
     </servlet-mapping>
     
     <servlet>
@@ -33,16 +33,16 @@
 
     <servlet>
         <servlet-name>add</servlet-name>
-        <servlet-class>online.shop.controller.ProductAddController</servlet-class>
+        <servlet-class>online.shop.controller.AddProductController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>add</servlet-name>
-        <url-pattern>/add</url-pattern>
+        <url-pattern>/products/add</url-pattern>
     </servlet-mapping>
 
     <servlet>
         <servlet-name>displayShoppingCart</servlet-name>
-        <servlet-class>online.shop.controller.ShoppingCartDisplayController</servlet-class>
+        <servlet-class>online.shop.controller.DisplayShoppingCartController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>displayShoppingCart</servlet-name>
@@ -55,12 +55,12 @@
     </servlet>
     <servlet-mapping>
         <servlet-name>addProductToCart</servlet-name>
-        <url-pattern>/shopping-cart/add</url-pattern>
+        <url-pattern>/shopping-carts/products/add</url-pattern>
     </servlet-mapping>
 
     <servlet>
         <servlet-name>allUsers</servlet-name>
-        <servlet-class>online.shop.controller.UsersController</servlet-class>
+        <servlet-class>online.shop.controller.GetAllUsersController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>allUsers</servlet-name>
@@ -69,7 +69,7 @@
 
     <servlet>
         <servlet-name>allProduct</servlet-name>
-        <servlet-class>online.shop.controller.ProductGetAllController</servlet-class>
+        <servlet-class>online.shop.controller.GetAllProductsController</servlet-class>
     </servlet>
     <servlet-mapping>
         <servlet-name>allProduct</servlet-name>
